+++ src/chanjo2/meta/handle_coverage_stats.py
@@ -10,14 +10,28 @@ STOP_INDEX = 2
 STATS_MEAN_COVERAGE_INDEX = 3
 
 
+def get_chromosomes_prefix(d4_file_path: str) -> str:
+    """Extracts the prefix to be prepended to genomic intervals when calculating stats.
+    Depends on the chromosomes format present on the d4 file..
+    """
+    chr1_stats = subprocess.check_output(
+        f"d4tools view -g {d4_file_path} | head -n 1", text=True, shell=True
+    )
+    if "chr" in chr1_stats:
+        return "chr"
+    return ""
+
+
 def get_d4tools_intervals_mean_coverage(
     d4_file_path: str, interval_ids_coords: List[Tuple[str, tuple]]
 ) -> List[float]:
     """Return the mean value over a list of intervals of a d4 file."""
 
+    chrom_prefix: str = get_chromosomes_prefix(d4_file_path)
+
     if interval_ids_coords:
         bed_lines = [
-            f"{coords[CHROM_INDEX]}\t{coords[START_INDEX]}\t{coords[STOP_INDEX]}"
+            f"{chrom_prefix}{coords[CHROM_INDEX]}\t{coords[START_INDEX]}\t{coords[STOP_INDEX]}"
             for _, coords in interval_ids_coords
         ]
         # Write genomic intervals to a temporary file
@@ -31,6 +45,7 @@ def get_d4tools_intervals_mean_coverage(
     chromosomes_mean_cov = get_d4tools_chromosome_mean_coverage(
         d4_file_path=d4_file_path, chromosomes=CHROMOSOMES
     )
+
     return [chrom_cov[1] for chrom_cov in chromosomes_mean_cov]
 
 
@@ -38,6 +53,9 @@ def get_d4tools_intervals_coverage(
     d4_file_path: str, bed_file_path: str
 ) -> List[float]:
     """Return the coverage for intervals of a d4 file that are found in a bed file."""
+
+    with open(bed_file_path, "r") as f:
+        print(f.read())
 
     d4tools_stats_mean_cmd: str = subprocess.check_output(
         ["d4tools", "stat", "--region", bed_file_path, d4_file_path, "--stat", "mean"],