+++ src/chanjo2/meta/handle_d4.py
@@ -1,4 +1,5 @@
 import logging
+import subprocess
 import tempfile
 from statistics import mean
 from typing import Dict, List, Optional, Tuple, Union