+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [unreleased]
+### Fixed
+- `/coverage/d4/genes/summary` endpoint is now returning stats for d4file with chromosomes containing the `chr` suffix
+
 ## [3.2]
 ### Added
 - `/intervals/intervals_count_by_build` endpoint, which retuns the number of genes, transcripts, exons for each genome build